Rebase and other pr changes
